+++ include/lex/token.hpp
@@ -7,7 +7,6 @@
 #include <iostream>
 #include <memory>
 #include <string>
-#include <utility>
 
 namespace lex {
