+++ src/CMakeLists.txt
@@ -6,7 +6,7 @@ add_library(lex STATIC
   lex/source_span.cpp
   lex/token.cpp
   lex/token_cat.cpp
-  lex/token_payload.cpp
+  lex/token_payload_lit_string.cpp
   lex/token_type.cpp
   lex/keyword.cpp
   lex/utilities.cpp)
