+++ tests/lex/keyword_test.cpp
@@ -1,6 +1,5 @@
 #include "catch2/catch.hpp"
 #include "helpers.hpp"
-#include <string>
 
 namespace lex {
