+++ tests/lex/operators_test.cpp
@@ -1,7 +1,5 @@
 #include "catch2/catch.hpp"
 #include "helpers.hpp"
-#include "lex/error.hpp"
-#include <string>
 
 namespace lex {
