+++ tests/parse/expr_call_test.cpp
@@ -1,5 +1,6 @@
 #include "catch2/catch.hpp"
 #include "helpers.hpp"
+#include "parse/error.hpp"
 #include "parse/node_expr_call.hpp"
 #include "parse/node_expr_group.hpp"
