+++ tests/parse/expr_paren_test.cpp
@@ -1,5 +1,6 @@
 #include "catch2/catch.hpp"
 #include "helpers.hpp"
+#include "parse/error.hpp"
 
 namespace parse {
