+++ tests/parse/expr_primary_test.cpp
@@ -1,5 +1,7 @@
 #include "catch2/catch.hpp"
 #include "helpers.hpp"
+#include "lex/error.hpp"
+#include "parse/error.hpp"
 
 namespace parse {
