+++ tests/parse/iterator_test.cpp
@@ -2,7 +2,6 @@
 #include "lex/lexer.hpp"
 #include "parse/parser.hpp"
 #include <string>
-#include <vector>
 
 namespace parse {
