+++ tests/parse/stmt_func_decl_test.cpp
@@ -2,10 +2,7 @@
 #include "helpers.hpp"
 #include "parse/error.hpp"
 #include "parse/node_expr_binary.hpp"
-#include "parse/node_expr_call.hpp"
-#include "parse/node_expr_group.hpp"
 #include "parse/node_stmt_func_decl.hpp"
-#include <utility>
 
 namespace parse {
